// src/app/Model/providers.ts
export interface Providers {

    // constructor( providerName:string, health:string, dental:string, vision: string, life: string){
    //     this.providerName = providerName;
    //     this.health = health;
    //     this.dental = dental;
    //     this.vision = vision;
    //     this.life = life;
    // }
    providerName: string;
    health: string;
    dental: string;
    vision: string;
    life: string;
}